# swagger/swagger.yaml
openapi: 3.0.3
info:
  title: KNFAPP API
  description: Backend API for VU Kaunas Faculty mobile app
  version: 1.7.0
  contact:
    name: VU KNF IT
servers:
  - url: /api
    description: API server
...
paths:
...
  /schedule:
    get:
      tags: [schedule]
      summary: Get schedule/timetable
      parameters:
        - name: day
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 6
        - name: group
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lessons for the day
          content:
            application/json:
              schema:
                type: object
                properties:
                  lessons:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lesson"
...
  /scraper/schedule:
    post:
      tags: [scraper]
      summary: Manually trigger schedule scrape (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Schedule scrape results
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Lesson:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        teacher:
          type: string
        room:
          type: string
        timeStart:
          type: string
        timeEnd:
          type: string
        dayOfWeek:
          type: integer
        group:
          type: string
        semester:
          type: string
...